LayoutTests/fast/forms/week/week-stepup-stepdown.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepUp() and stepDown() behavior for type=week.');

var input = document.createElement('input');

function setInputAttributes(min, max, step, value) {
    input.min = min;
    input.max = max;
    input.step = step;
    input.value = value;
}

function stepUp(value, step, max, optionalStepCount) {
    setInputAttributes(null, max, step, value);
    if (typeof optionalStepCount != "undefined")
        input.stepUp(optionalStepCount);
    else
        input.stepUp();
    return input.value;
}

function stepDown(value, step, min, optionalStepCount) {
    setInputAttributes(min, null, step, value);
    if (typeof optionalStepCount != "undefined")
        input.stepDown(optionalStepCount);
    else
        input.stepDown();
    return input.value;
}

input.type = 'week';
debug('Invalid value');
shouldBeEqualToString('stepUp("", null, null)', '1970-W02');
shouldBeEqualToString('stepDown("", null, null)', '1970-W01');
debug('Non-number arguments');
shouldBeEqualToString('stepUp("2010-W02", null, null, "0")', '2010-W02');
shouldBeEqualToString('stepDown("2010-W02", null, null, "0")', '2010-W02');
shouldBeEqualToString('stepUp("2010-W02", null, null, "foo")', '2010-W02');
shouldBeEqualToString('stepDown("2010-W02", null, null, "foo")', '2010-W02');
shouldBeEqualToString('stepUp("2010-W02", null, null, null)', '2010-W02');
shouldBeEqualToString('stepDown("2010-W02", null, null, null)', '2010-W02');
debug('Normal cases');
shouldBeEqualToString('stepUp("2010-W02", null, null)', '2010-W03');
shouldBeEqualToString('stepDown("2010-W02", null, null)', '2010-W01');
shouldBeEqualToString('stepUp("2010-W02", null, null, 10)', '2010-W12');
shouldBeEqualToString('stepDown("2010-W02", null, null, 11)', '2009-W44');
shouldBeEqualToString('stepUp("1970-W01", "4", null, 2)', '1970-W09');
shouldBeEqualToString('stepDown("1970-W01", "4", null, 3)', '1969-W41');
debug('Step=any');
shouldThrow('stepUp("2010-W02", "any", null)');
shouldThrow('stepDown("2010-W02", "any", null)');
debug('Overflow/underflow');
shouldBeEqualToString('stepUp("2010-W02", "3.40282346e+38", null)', '1970-W01');
shouldBeEqualToString('stepDown("2010-W02", "3.40282346e+38", null)', '1970-W01');
shouldBeEqualToString('stepUp("2010-W02", "1", "2010-W02")', '2010-W02');
shouldBeEqualToString('stepDown("2010-W02", "1", "2010-W02")', '2010-W02');

debug('');
</script>
</body>
</html>
